Removing unused code
